Add maxmind endpoint

+++ models/geolocation.go
@@ -1,9 +1,9 @@
 package models
 
 type Geolocation struct {
-	IPAddress string	`json:"ipaddress"`
-	City string      	`json:"city"`
-	Country string   	`json:"country"`
-	Latitude float32 	`json:"latitude"`
-	Longitude float32	`json:"longitude"`
+	IPAddress string  `json:"ipaddress"`
+	City      string  `json:"city"`
+	Country   string  `json:"country"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
